[35.1] 다른 사용자 프로필 확인 UI & 로직 제작
- API사양 변경 : rating = double -> Int

+++ data/src/main/kotlin/kr/linkerbell/campusmarket/android/data/remote/network/model/feature/mypage/UserReviewRes.kt
@@ -31,7 +31,7 @@ data class ReviewItemRes(
     @SerialName("description")
     val description: String,
     @SerialName("rating")
-    val rating: Double,
+    val rating: Int,
     @SerialName("createdAt")
     val createdAt: LocalDateTime
 ) : DataMapper<UserReview> {
@@ -41,7 +41,7 @@ data class ReviewItemRes(
             nickname = nickname,
             profileImage = profileImage,
             description = description,
-            rating = rating.toInt(),
+            rating = rating,
             createdAt = createdAt
         )
     }
